refactor(cli): commands/{eval,observe}.ts → commands/{eval,observe}/index.ts
视觉收口:之前 `eval.ts` 跟 `eval/` 在 commands 目录里并排,IDE 列表
file 跟同名 directory 挨在一起;observe 同情况。挪进同名 directory 作
default command(oclif 4 接受 `<topic>/index.ts` 作 topic 默认命令)
后,commands 目录变成「扁平的纯 file、有 sub 的纯 directory」一致风格:

```
commands/
  doctor.ts        # 扁平命令
  evolve.ts
  init.ts
  sample.ts
  studio.ts
  eval/            # 有 sub 的命令
    index.ts       # omk eval default
    gold.ts        # eval gold topic shim
    gold/
      init.ts validate.ts compare.ts
  observe/
    index.ts       # omk observe default
    ingest.ts inbox.ts show.ts
```

仅文件位置变,index.ts 内部业务跟 oclif Command class 完全不动;两个
file 内的 relative import 按新 depth +1 重写(static + dynamic 一致)。
CONTRIBUTING.md 文件目录约定同步更新,标 oclif file routing 约定。

smoke test 验证 oclif 把 index.ts 正确识别为 default command:
- `omk eval --help` → eval 自身 help(不是 root help)
- `omk eval gold --help` → topic shim(不是 eval 自身)
- `omk observe inbox --help` → sub command
- `omk eval --dry-run --control baseline --treatment x` → 业务路径(走到
  缺 sample 文件,正常)

yarn lint && build && build:docs:check && test 1498 全绿。
